refactor: remove unused proposal components and update imports
