+++ src/app/proposals/new/components/CardParagraph.tsx
