+++ src/app/proposals/new/components/Deactivation.tsx
