+++ src/app/proposals/new/components/NewProposalCard.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
-import { type NewProposalCardBaseData } from './newProposalCards.data'
-import { DotsOverlayVert } from './images/DotsOverlayVert'
+import { type NewProposalCardBaseData } from '../newProposalCards.data'
+import { DotsOverlayVert } from '../images/DotsOverlayVert'
 import { CardButton } from './CardButton'
 import { HTMLAttributes } from 'react'
 import { cn } from '@/lib/utils'
