+++ src/app/proposals/new/components/NewProposalCardExtended.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { NewProposalCardExtended } from './NewProposalCardExtended'
-import { newProposalCards, type NewProposalCardExtendedData } from './newProposalCards.data'
+import { newProposalCards, type NewProposalCardExtendedData } from '../newProposalCards.data'
 import { useState } from 'react'
 import { AnimatePresence } from 'motion/react'
 import { Button } from '@/components/ButtonNew'
