+++ src/app/proposals/new/page.tsx
@@ -3,11 +3,11 @@
 import { useCallback, useEffect, useMemo } from 'react'
 import { AnimatePresence, motion, Variants } from 'motion/react'
 import { useRouter, useSearchParams, usePathname } from 'next/navigation'
-import { Deactivation } from './Deactivation'
-import { NewProposalCard } from './NewProposalCard'
+import { Deactivation } from './components/Deactivation'
+import { NewProposalCard } from './components/NewProposalCard'
 import { newProposalCards } from './newProposalCards.data'
 import { Paragraph } from '@/components/TypographyNew'
-import { NewProposalCardExtended } from './NewProposalCardExtended'
+import { NewProposalCardExtended } from './components/NewProposalCardExtended'
 import { ProposalCategory } from '@/shared/types'
 import { useProposalStepper } from './stepper/StepperProvider'
